Add a method to merge multiple user timelines
This approach is not really documented by mixpanel but
uses some of its internals to merge multiple user timelines.
It works well with our proxy that maps these calls to appropriate Amplitude API
to associate multiple device IDs with a single user name.

On mixpanel side, this results in users duplication when
multiple device IDs are used

Change-type: minor
Signed-off-by: Roman Mazur <mazur.roman@gmail.com>